+++ tailwind.config.js
@@ -1,6 +1,7 @@
 export default {
     content: [
         './resources/js/**/*.{js,ts,vue}',
+        './node_modules/atlas-ui/src/**/*.{js,ts,vue}',
     ],
     plugins: [
         require('@tailwindcss/forms'),